Move api and model to data package
