+++ app/src/main/java/me/pedroguimaraes/weatherdemo/data/model/Weather.kt
@@ -1,4 +1,4 @@
-package me.pedroguimaraes.weatherdemo.model
+package me.pedroguimaraes.weatherdemo.data.model
 
 data class Weather(val latitude: Double, val longitude: Double, val timezone: String,
 				   val offset: Int, val currently: Currently)
