+++ app/src/main/java/me/pedroguimaraes/weatherdemo/ui/weather/WeatherActivity.kt
@@ -7,9 +7,9 @@ import android.support.v7.app.AppCompatActivity
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.layout_weather_today.*
 import me.pedroguimaraes.weatherdemo.R
+import me.pedroguimaraes.weatherdemo.data.model.WeatherInfo
 import me.pedroguimaraes.weatherdemo.injection.DependencyInjection
 import me.pedroguimaraes.weatherdemo.interactors.permissions.PermissionEnforcer
-import me.pedroguimaraes.weatherdemo.model.WeatherInfo
 
 class WeatherActivity : AppCompatActivity(), WeatherContract.WeatherView {
