+++ app/src/test/java/me/pedroguimaraes/weatherdemo/data/model/WeatherInfoTest.kt
@@ -1,4 +1,4 @@
-package me.pedroguimaraes.weatherdemo.model
+package me.pedroguimaraes.weatherdemo.data.model
 
 import me.pedroguimaraes.weatherdemo.R
 import kotlin.test.assertEquals
